+++ src/framework/piral-core/src/actions/components.ts
@@ -1,46 +1,18 @@
-import { appendItem, excludeOn, withKey, withoutKey } from '../utils';
+import { withExtension, withoutExtension, withoutPage, withPage } from '../utils';
 import { PageRegistration, ExtensionRegistration, GlobalStateContext } from '../types';
 
 export function registerPage(ctx: GlobalStateContext, name: string, value: PageRegistration) {
-  ctx.dispatch((state) => ({
-    ...state,
-    registry: {
-      ...state.registry,
-      pages: withKey(state.registry.pages, name, value),
-    },
-  }));
+  ctx.dispatch(withPage(name, value));
 }
 
 export function unregisterPage(ctx: GlobalStateContext, name: string) {
-  ctx.dispatch((state) => ({
-    ...state,
-    registry: {
-      ...state.registry,
-      pages: withoutKey(state.registry.pages, name),
-    },
-  }));
+  ctx.dispatch(withoutPage(name));
 }
 
 export function registerExtension(ctx: GlobalStateContext, name: string, value: ExtensionRegistration) {
-  ctx.dispatch((state) => ({
-    ...state,
-    registry: {
-      ...state.registry,
-      extensions: withKey(state.registry.extensions, name, appendItem(state.registry.extensions[name], value)),
-    },
-  }));
+  ctx.dispatch(withExtension(name, value));
 }
 
 export function unregisterExtension(ctx: GlobalStateContext, name: string, reference: any) {
-  ctx.dispatch((state) => ({
-    ...state,
-    registry: {
-      ...state.registry,
-      extensions: withKey(
-        state.registry.extensions,
-        name,
-        excludeOn(state.registry.extensions[name], (m) => m.reference === reference),
-      ),
-    },
-  }));
+  ctx.dispatch(withoutExtension(name, reference));
 }