+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Piral Changelog
 
+## 0.14.4 (tbd)
+
+- Improved siteless API to support using other frameworks
+- Fixed issue with entry point of pilet missing
+- Added extensions for each plugin component (e.g., `piral-dashboard` for the `Dashboard` component of `piral-dashbard`)
+
 ## 0.14.3 (November 26, 2021)
 
 - Improved `pilet-uses-latest-piral` rule w.r.t. monorepos